Update block imports
