+++ src/client/terrain/util/mesher.rs
@@ -1,11 +1,4 @@
-use super::{
-    blocks::{Block, BlockId},
-    chunk::{Chunk, CHUNK_SIZE},
-};
-use bevy::render::{
-    mesh::{Indices, Mesh, PrimitiveTopology},
-    render_asset::RenderAssetUsages,
-};
+use crate::prelude::*;
 
 pub fn create_cube_mesh_from_data(geometry_data: GeometryData) -> Option<Mesh> {
     let GeometryData {
